Dependencies grouping, view component, fixed repo

+++ HistoryPediaV2.Core/Repositories/IGenericRepository.cs
@@ -3,14 +3,14 @@
     public interface IGenericRepository<T> where T : class
     {
         // Task<T> GetAsync(int? id);
-        Task<TResult> GetAsync<TResult>(int? id);
+        Task<TResult> GetAsync<TResult>(long? id);
         // Task<List<T>> GetAllAsync();
         Task<List<TResult>> GetAllAsync<TResult>();
         // Task<T> AddAsync(T entity);
         Task<TResult> AddAsync<TSource, TResult>(TSource source);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
         // Task UpdateAsync(T entity);
-        Task UpdateAsync<TSource>(int id, TSource source);
+        Task UpdateAsync<TSource>(long id, TSource source);
         // Task<bool> Exists(int id);
     }
 }
